[Spring Security] 보안 코드 개선 (#113)
* SpringSecurity & JWT

* application.yml - jwt information

* MemberDetailsService 오류사항 수정

* AuthenticationSuccessHandler, AuthenticationFailureHandler 구현

* JwtVerificationFilter 구현 및 member 앤티티의 username 속성 수정(unique 삭제)

* 회원가입 성공 시, 로그 출력 & 개발팀원 이메일 admin권한 등록

* security 오류 수정

* 코드(member entity, SecurityConfiguration) 및 주석(JwtVerificationFilter) 수정

* QuestionService에 verifyAuthorizedMember() 추가 및 에러코드 추가(MEMBER_NOT_VALID)

* AnswerService에 verifyAuthorizedMember() 추가

* CommentService에 verifyAuthorizedMember() 추가

* MemberService에 verifyAuthorizedMember() 추가

* springSecurity 권한 matcher 수정(전체조회 시, 엔드포인트에 "/" 추가하면 관리자가 아닌 일반 회원도 리소스 접근 가능한 취약점 수정)

* 관리자(admin) 이메일을 verifyAuthorizedMember() 메서드에 추가.

* 특정회원정보조회는 회원가입한 회원 본인만 가능하도록 수정.

* Answer_Selected 는 해당 답변이 종속되어있는 질문의 작성자만 수정가능토록 코드 개선 & 주석 수정

+++ backend/preproject/src/main/java/com/seb43_pre_12/preproject/answers/service/AnswerService.java
@@ -5,12 +5,16 @@ import com.seb43_pre_12.preproject.exception.BusinessLogicException;
 import com.seb43_pre_12.preproject.answers.repository.AnswerRepository;
 import com.seb43_pre_12.preproject.exception.ExceptionCode;
 import com.seb43_pre_12.preproject.member.entity.Member;
+import com.seb43_pre_12.preproject.member.repositoy.MemberRepository;
 import com.seb43_pre_12.preproject.member.service.MemberService;
 import com.seb43_pre_12.preproject.question.entity.Question;
 import com.seb43_pre_12.preproject.question.service.QuestionService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -22,10 +26,18 @@ public class AnswerService {
     private final AnswerRepository repository;
     private final MemberService memberService;
     private final QuestionService questionService;
-    public AnswerService(AnswerRepository repository, MemberService memberService, QuestionService questionService) {
+    private final MemberRepository memberRepository;
+//    public AnswerService(AnswerRepository repository, MemberService memberService, QuestionService questionService) {
+//        this.repository = repository;
+//        this.memberService = memberService;
+//        this.questionService = questionService;
+//    }
+
+    public AnswerService(AnswerRepository repository, MemberService memberService, QuestionService questionService, MemberRepository memberRepository) {
         this.repository = repository;
         this.memberService = memberService;
         this.questionService = questionService;
+        this.memberRepository = memberRepository;
     }
 
     public Answer createAnswer(Answer answer) {
@@ -41,12 +53,30 @@ public class AnswerService {
         return repository.save(answer);
     }
     public Answer updateAnswer(Answer answer) {
+        String loginEmail = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String questionOwnerEmail = findVerifedAnswer(answer.getAnswerId()).getQuestion().getMember().getEmail();
+        String answerOwnerEmail = findVerifedAnswer(answer.getAnswerId()).getMember().getEmail();
+        List<String> adminMailAddress = List.of("hw@email.com", "ny@email.com","sh@email.com","hj@email.com","jh@email.com","jm@email.com");
+
         Answer findAnswer = findVerifedAnswer(answer.getAnswerId());
 
-        Optional.ofNullable(answer.getContent()).ifPresent(content -> findAnswer.setContent(content));
-        Optional.ofNullable(answer.getSelected()).ifPresent(selected -> findAnswer.setSelected(selected));
-        Optional.ofNullable(answer.getMember()).ifPresent(member -> findAnswer.setMember(member));
-        Optional.ofNullable(answer.getQuestion()).ifPresent(question -> findAnswer.setQuestion(question));
+        if( adminMailAddress.contains(loginEmail)) {
+            Optional.ofNullable(answer.getContent()).ifPresent(content -> findAnswer.setContent(content)); // 관리자, 답변작성자 만 수정가능
+            Optional.ofNullable(answer.getSelected()).ifPresent(selected -> findAnswer.setSelected(selected));// 관리자, 질문작성자 만 수정가능
+        }
+
+        if(loginEmail.equals(answerOwnerEmail) || loginEmail.equals(questionOwnerEmail)) {
+            if(loginEmail.equals(answerOwnerEmail)) {
+                Optional.ofNullable(answer.getContent()).ifPresent(content -> findAnswer.setContent(content)); // 관리자, 답변작성자 만 수정가능
+            }
+            if (loginEmail.equals(questionOwnerEmail)) {
+                Optional.ofNullable(answer.getSelected()).ifPresent(selected -> findAnswer.setSelected(selected));// 관리자, 질문작성자 만 수정가능
+            }
+        } else if (adminMailAddress.contains(loginEmail)) {}
+        else throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_VALID);
+
+//        Optional.ofNullable(answer.getMember()).ifPresent(member -> findAnswer.setMember(member));
+//        Optional.ofNullable(answer.getQuestion()).ifPresent(question -> findAnswer.setQuestion(question));
 
         findAnswer.setModifiedAt(LocalDateTime.now());
 
@@ -64,6 +94,7 @@ public class AnswerService {
 
 
     public void deleteAnswer(long answerId) {
+        verifyAuthorizedMember(answerId);
         repository.delete(findVerifedAnswer(answerId));
     }
 
@@ -81,5 +112,19 @@ public class AnswerService {
 
     private Member verifyExistingMember(Member member) {
         return memberService.findVerifiedMember(member.getMemberId());
+    }
+    private void verifyAuthorizedMember(Long answerId) {
+        // 현재 로그인한 회원의 이메일을 찾는 로직
+        String loginEmail = (String)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        // 댭변을 작성한 회원 객체를 찾는 로직
+        String AnswerOwnerEmail = findVerifedAnswer(answerId).getMember().getEmail();
+        // 관리자 계정 리스
+        List<String> adminMailAddress = List.of("hw@email.com", "ny@email.com","sh@email.com","hj@email.com","jh@email.com","jm@email.com");
+
+        // 댭변을 작성한 회원 객체의 email 과 로그인한 회원의 email 이 동일한지 조건문을 통해서 검사한다.
+        if(adminMailAddress.contains(loginEmail)) return;
+        else if (loginEmail.equals(AnswerOwnerEmail)) return;
+        else throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_VALID);
     }
 }
