+++ backend/preproject/src/main/java/com/seb43_pre_12/preproject/member/service/MemberService.java
@@ -1,5 +1,6 @@
 package com.seb43_pre_12.preproject.member.service;
 
+import com.seb43_pre_12.preproject.answers.entity.Answer;
 import com.seb43_pre_12.preproject.auth.CustomAuthorityUtils;
 import com.seb43_pre_12.preproject.exception.BusinessLogicException;
 import com.seb43_pre_12.preproject.exception.ExceptionCode;
@@ -7,6 +8,9 @@ import com.seb43_pre_12.preproject.exception.ExceptionCode;
 import com.seb43_pre_12.preproject.member.entity.Member;
 import com.seb43_pre_12.preproject.member.repositoy.MemberRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -54,6 +58,7 @@ public class MemberService {
     }
 
     public Member updateMember(Member member) {
+        verifyAuthorizedMember(member.getMemberId());
         Member findMember = findVerifiedMember(member.getMemberId());
 
         Optional.ofNullable(member.getEmail())
@@ -71,6 +76,7 @@ public class MemberService {
 
     @Transactional(readOnly = true)
     public Member findMember(long memberId) {
+        verifyAuthorizedMember(memberId);
         Member findMember = findVerifiedMember(memberId);
 
 //         comments 출력
@@ -84,6 +90,7 @@ public class MemberService {
     }
 
     public void deleteMember(long memberId) {
+        verifyAuthorizedMember(memberId);
         Member findMember = findVerifiedMember(memberId);
 
         memberRepository.delete(findMember);
@@ -101,5 +108,19 @@ public class MemberService {
         if (member.isPresent())
             throw new BusinessLogicException(ExceptionCode.MEMBER_EXISTS);
     }
+    private void verifyAuthorizedMember(Long memberId) {
+        // 현재 로그인한 회원의 이메일을 찾는 로직트
+        String loginEmail = (String)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
+        // 회원가입을 진행한 실제 회원 객체를 찾는 로직
+        final String ownerMemberEmail = findVerifiedMember(memberId).getEmail();
+        // 관리자 계정 리스트
+        List<String> adminMailAddress = List.of("hw@email.com", "ny@email.com","sh@email.com","hj@email.com","jh@email.com","jm@email.com");
+
+        // 회원가입을 진행한 실제 회원 객체의 email 과 로그인한 회원의 email 이 동일한지 조건문을 통해서 검사한다.
+        if(adminMailAddress.contains(loginEmail)) return;
+        else if (loginEmail.equals(ownerMemberEmail)) return;
+        else throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_VALID);
+
+    }
 }
