+++ backend/preproject/src/main/java/com/seb43_pre_12/preproject/question/service/QuestionService.java
@@ -1,19 +1,27 @@
 package com.seb43_pre_12.preproject.question.service;
 
+import com.seb43_pre_12.preproject.auth.JwtTokenizer;
 import com.seb43_pre_12.preproject.exception.BusinessLogicException;
 import com.seb43_pre_12.preproject.exception.ExceptionCode;
 import com.seb43_pre_12.preproject.member.entity.Member;
+import com.seb43_pre_12.preproject.member.repositoy.MemberRepository;
 import com.seb43_pre_12.preproject.member.service.MemberService;
 import com.seb43_pre_12.preproject.question.entity.Question;
 import com.seb43_pre_12.preproject.question.repository.QuestionRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
@@ -41,6 +49,8 @@ public class QuestionService {
     }
     @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
     public Question updateQuestion(Question question){
+        verifyAuthorizedMember(question.getQuestionId());
+
         Question findQuestion = findVerifiedQuestion(question.getQuestionId());
 
         Optional.ofNullable(question.getContent())
@@ -65,6 +75,8 @@ public class QuestionService {
                 Sort.by("createdAt").descending()));
     }
     public void deleteQuestion(Long questionId){
+        verifyAuthorizedMember(questionId);
+
         Question findQuestion = findVerifiedQuestion(questionId);
 
         questionRepository.delete(findQuestion);
@@ -84,4 +96,19 @@ public class QuestionService {
     private Member verifyExistingMember(Member member){
         return memberService.findVerifiedMember(member.getMemberId());
     }
-}
+
+    private void verifyAuthorizedMember(Long questionId) {
+        // 현재 로그인한 회원의 이메일을 찾는 로직
+        String loginEmail = (String)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        // 질문을 작성한 회원 객체를 찾는 로직
+        final String questionOwnerEmail = findVerifiedQuestion(questionId).getMember().getEmail();
+        // 관리자 계정 리스트
+        List<String> adminMailAddress = List.of("hw@email.com", "ny@email.com","sh@email.com","hj@email.com","jh@email.com","jm@email.com");
+
+        // 질문을 작성한 회원 객체의 email 과 로그인한 회원의 email 이 동일한지 조건문을 통해서 검사한다.
+        if(adminMailAddress.contains(loginEmail)) return;
+        else if (loginEmail.equals(questionOwnerEmail)) return;
+        else throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_VALID);
+        }
+    }
